Get color

+++ image_processing_module/matrix_color_service.py
@@ -158,7 +158,7 @@ class MatrixColorService:
                 break
         return main_different_colors
     
-    def change_gamma_colors(self, color: List[int], new_color: List[int], delta_threshold: float = 3.0) -> NDArray[np.float64]:
+    def change_gamma_colors(self, color: List[float], new_color: List[float], delta_threshold: float = 3.0) -> NDArray[np.float64]:
         color_arr = np.array(color, dtype=np.float64)
         new_color_arr = np.array(new_color, dtype=np.float64)
         delta_arr = new_color_arr - color_arr
